fix: organize lib
